A Balanced Ternary Clock
This is an Android Wear Watchface for a balanced ternary clock. It tells
the hours past noon, minutes past the hour and seconds past the minute.
The date is days of the week past Monday, day of the month, month of the
year, and year. The watch face adjusts for peek card size. This should
have an alarm manager to update on the half minutes in ambient mode. And
it should have a configuration activity to select how to display the
date.